Add files via upload
Program the return the minimum questions to solve to achieve the given threshold value. the questions are given in the form of array where each number represents the marks for that particular question. Threhold is calculated as max mark que - min mark question.
eg. Array - [1,2,3,4]. threshold - 3. o/p = [1,2,4] or [1,3,4] i.e 3 questions need to be solved to achieve threshold(4-1=3)
NOTE - a person must solve first question and then can skip at the most 1 question to proceed. Also the array given is sorted.